+++ react-app/src/components/Main/Chat/Inputs/InputChat.js
@@ -4,29 +4,27 @@ import { useParams } from "react-router-dom";
 
 import { socket } from "../../../../context/Socket";
 
-const InputChat = () => {
+const InputChat = ({ chat, setChat, handleSubmit }) => {
 	const channelId = parseInt(useParams().channelId, 10);
 	const channels = useSelector((state) => state.channels);
 	const channelsArr = channels.byId;
 	const channel = useSelector((state) => state.channels.byId)[channelId];
 
-	const [chat, setChat] = useState("");
+	// const handleSubmit = async (e) => {
+	// 	e.preventDefault();
+	// 	// only send when there is chat input
+	// 	if (chat) {
+	// 		const chatData = {
+	// 			channel_id: channelId,
+	// 			chat,
+	// 		};
 
-	const handleSubmit = async (e) => {
-		e.preventDefault();
-		// only send when there is chat input
-		if (chat) {
-			const chatData = {
-				channel_id: channelId,
-				chat,
-			};
+	// 		socket.emit("send_chat", chatData);
 
-			socket.emit("send_chat", chatData);
-
-			// clear chat field
-			setChat("");
-		}
-	};
+	// 		// clear chat field
+	// 		setChat("");
+	// 	}
+	// };
 
 	return (
 		<form className="chat-input-ctrl" onSubmit={handleSubmit}>
