+++ react-app/src/components/Main/index.js
@@ -1,6 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Switch, Route } from "react-router-dom";
+// import { io } from "socket.io-client";
 
 import SideBar from "./SideBar";
 import TopBar from "./TopBar";
@@ -18,7 +19,8 @@ import * as prvChatsActions from "../../store/prvChats";
 import * as joinServersActions from "../../store/joinServers";
 import * as usersActions from "../../store/users";
 
-import { socket } from "../../context/Socket";
+// import { socket } from "../../context/Socket";
+let socket;
 
 const MainPage = () => {
 	const dispatch = useDispatch();
@@ -40,25 +42,26 @@ const MainPage = () => {
 			})
 			.catch((err) => console.log(err));
 
-		socket.on("receive_message", (data) => {
-			// dispatch chat with id to our redux store
-			dispatch(channelsActions.addEditChannel(data.channel));
-			// dispatch(chatActions.addChat(data));
-			dispatch(chatsActions.addEditChat(data.chat));
-		});
+		// socket = io();
+
+		// socket.on("receive_message", (data) => {
+		// 	// dispatch chat with id to our redux store
+		// 	dispatch(channelsActions.addEditChannel(data.channel));
+		// 	// dispatch(chatActions.addChat(data));
+		// 	dispatch(chatsActions.addEditChat(data.chat));
+		// });
 
 		// Disconnect socket when leave page
 		return () => {
-			// socket.emit("leave_channels", channelArr);
-			socket.disconnect();
+			// socket.disconnect();
 		};
 	}, [dispatch]);
 
-	useEffect(() => {
-		setTimeout(() => {
-			socket.emit("join_channels", channelArr);
-		}, 5000);
-	}, [channelArr]);
+	// useEffect(() => {
+	// 	setTimeout(() => {
+	// 		socket.emit("join_channels", channelArr);
+	// 	}, 5000);
+	// }, [channelArr]);
 
 	return (
 		<div className="main-ctrl">
